+++ Source/AmazonPublisherServicesAdapterConfiguration.swift
@@ -32,4 +32,16 @@ import os.log
             }
         }
     }
+
+    /// A delegate that performs pre-bidding operations by integrating directly with the Amazon Publisher Services SDK.
+    /// Publishers are required to implement the `AmazonPublisherServicesAdapterPreBiddingDelegate` delegate and set this
+    /// property during app initialization.
+    ///
+    /// Chartboost is not permitted to wrap the Amazon APS initialization or bid request methods directly.
+    /// The adapter handles APS initialization and prebidding only when the managed prebidding flag is enabled.
+    /// For more information please contact the Amazon APS support team at https://aps.amazon.com/aps/contact-us/
+    @objc public static var preBiddingDelegate: AmazonPublisherServicesAdapterPreBiddingDelegate? {
+        get { AmazonPublisherServicesAdapter.preBiddingDelegate }
+        set { AmazonPublisherServicesAdapter.preBiddingDelegate = newValue }
+    }
 }